+++ src/app/loadout-drawer/LoadoutDrawerHeader.tsx
@@ -1,8 +1,11 @@
 import ClassIcon from 'app/dim-ui/ClassIcon';
+import { useAutocomplete } from 'app/dim-ui/text-complete/text-complete';
 import { t } from 'app/i18next-t';
-import React from 'react';
+import React, { useRef } from 'react';
+import { useSelector } from 'react-redux';
 import { Loadout } from './loadout-types';
 import styles from './LoadoutDrawerHeader.m.scss';
+import { loadoutsHashtagsSelector } from './selectors';
 
 export default function LoadoutDrawerHeader({
   loadout,
@@ -13,12 +16,17 @@ export default function LoadoutDrawerHeader({
 }) {
   const setName = (e: React.ChangeEvent<HTMLInputElement>) => onNameChanged(e.target.value);
 
+  const ref = useRef<HTMLInputElement>(null);
+  const tags = useSelector(loadoutsHashtagsSelector);
+  useAutocomplete(ref, tags);
+
   return (
     <div className={styles.loadoutName}>
       <ClassIcon classType={loadout.classType} />
       <input
         className={styles.dimInput}
         name="name"
+        ref={ref}
         onChange={setName}
         minLength={1}
         maxLength={50}